Update conference.md
Hi Nicolas, I applied some updates (information on the poster prizes amongst others) and I will send you 3 attachments to be added to the website: an update for the program, and 2 attachments for the poster sessions. I hope you still can do this as I understood that you are stepping down from the NVP board. Bests, Tildie